Now using std::error_code extensively.

+++ tests/url_web_platform_tests.cpp
@@ -110,7 +110,7 @@ TEST_P(test_parse_urls_using_base_urls, parse_using_make) {
   auto instance = skyr::make_url(test_case_data.input, skyr::url(test_case_data.base));
   ASSERT_TRUE(instance)
     << "Input: [" << test_case_data.input << "], Base: [" << test_case_data.base << "] "
-    << static_cast<int>(instance.error());
+    << instance.error();
   EXPECT_EQ(test_case_data.protocol, instance.value().protocol())
     << "Input: [" << test_case_data.input << "], Base: [" << test_case_data.base << "]";
   EXPECT_EQ(test_case_data.username, instance.value().username())
